refactor: restructure validation tests into focused modules
- Split chatmate_validation_test.go (309 lines) into focused test files:
  - test/validation/embedded_assets_test.go (131 lines): Embedded asset validation
  - test/validation/content_validation_test.go (135 lines): Content validation tests
  - test/validation/helper_functions_test.go (90 lines): Helper function tests
  - test/chatmate_validation_test.go (16 lines): Backward compatibility wrapper

- All files now comply with 300-line limit
- Maintains all existing test functionality
- Improves test organization by validation category
- Enables focused testing of specific validation aspects

File size compliance achieved:
✅ tutorial.go: 682→103 lines
✅ test_helpers.go: 432→72 lines (modular)
✅ integration_tests_test.go: 323→multiple <120 line files
✅ chatmate_validation_test.go: 309→16 lines (modular)
